bump the Poco version requirement to >= 1.5.0
With the 1.5 release of Poco, the Util library was changed to depend on the new
JSON library. Previously, building storyteller against Poco 1.5.x on a reasonably
recent Linux distribution would result in "unresolved symbol" complaints from
the linker regarding the JSON library.

This change bumps the version requirement to at least 1.5.0, and adds JSON to
the list of Poco components that get searched.